restructure project and switch to setuptools
Renamed the cli tools from bin/* and contrib/* to filenames without dashes so
they are loadable and easier to unittest. Moved those files to a directory in
the lib dir. Created a module subdirectory keep the pythonpath root dir clean.

Switch to setuptools so that scripts can be provided with `console_scripts`.