calculate offsets
To determine at which position in an archive to write a point Whisper would read
position one and then calculate the offset from that timestamp. Instead of
reading position one, now we calculate the offset directly saving a read
operation. This is approximately 20% faster according to a simple benchmark.

This is backwards incompatible, therefore updating to Whisper V2.